The last act of cursor
Pedro & MainDrive improvements, Panels toggle, auto-park, and cleanup.

- PedroDrive.getVelocity(): use Pedro Vector getXComponent/getYComponent and
  getAngularVelocity() so velocity-based turret/hood compensation works.

- Panels vs FTC Dashboard: MainDrive.usePanelsDashboard toggles telemetry
  to Panels (panels.bylazar.com) instead of FTC Dashboard; reflection-based
  so it works with FullPanels when available. Docs in PEDRO_MIGRATION.md.

- Auto-park: G1 left bumper toggles path-to-goal (parking spot). G1 right
  stick nudges target at small scale (parkNudgeScaleIn, parkNudgeDeadband).
  PedroDrive.breakFollowing() added to cancel path when exiting auto-park.
  Field overlay shows park target (purple) when active.

- MainDrive cleanup: class Javadoc (Drive, Auto-park, Turret, Alliance,
  Airsort), section comments in main loop, updateAutoPark() extracted.
  Dashboard packet and telemetry include auto_park and park nudge.

- Repo state: Pedro pathing (PedroDrive, pedroPathing.Constants), util
  (Pose2d, Vector2d, Actions, FollowPathAction, etc.), build/docs/scripts,
  autos and RR updates, removed unused subsystems (Elevator, Pivot, etc.).

Made-with: Cursor

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RR/PoseBridge.java
@@ -1,15 +1,15 @@
 package org.firstinspires.ftc.teamcode.RR;
 
-import com.acmerobotics.roadrunner.Pose2d;
+import org.firstinspires.ftc.teamcode.util.Pose2d;
 
 /**
  * Bridges field position, heading, and alliance from Autonomous to Teleop.
  * Static state persists while the app is running, so when you run Auto then Teleop
  * in the same session, Teleop can start from the pose Auto ended at and the same alliance.
  *
- * <p>Auto: call {@link #save(Pose2d)} with the current localizer pose (e.g. at end of autonomous
+ * <p>Auto: call {@link #save(Pose2d)} with the current pose (e.g. at end of autonomous
  * or each loop). Call {@link #saveAlliance(boolean)} with true for Blue, false for Red.
- * Teleop: in init, use {@link #getPose()} and pass to your localizer's setPose
+ * Teleop: in init, use {@link #getPose()} and pass to your drive's setPose
  * if {@link #hasPose()} is true. Use {@link #getAlliance()} if {@link #hasAlliance()} to set
  * the alliance (true = Blue, false = Red).
  */
@@ -27,7 +27,7 @@ public final class PoseBridge {
     public static void save(Pose2d pose) {
         x = pose.position.x;
         y = pose.position.y;
-        headingRad = pose.heading.toDouble();
+        headingRad = pose.heading;
         hasPose = true;
     }
